+++ src/app/services/currencies.service.ts
@@ -1,6 +1,7 @@
 import { Injectable } from '@angular/core';
-import { environment } from '../../environments/environment';
-import { Utils } from './../utils/utils';
+import { Utils } from 'app/utils/utils';
+import { Currency } from 'app/utils/currencies';
+import { ChainService } from 'app/services/chain.service';
 
 interface BestInterestRate {
   min: number;
@@ -27,35 +28,15 @@ export class CurrenciesService {
     'USDC': { min: 5, max: 10, best: 7 },
     'ARS': { min: 40, max: 80, best: 60 },
     'ETH': { min: 0, max: 5, best: 1 },
+    'BNB': { min: 0, max: 5, best: 1 },
     'TEST': { min: 0, max: 20, best: 10 },
     'DEST': { min: 10, max: 20, best: 15 }
   };
 
-  constructor() {
-    const envCurrencies = environment.usableCurrencies;
-    const currencies: CurrencyItem[] = [];
-
-    envCurrencies.map((currency: {
-      symbol: string;
-      address: string;
-      img: string;
-    }) => {
-      const address = currency.address;
-      const symbol = currency.symbol;
-      const img = currency.img;
-      const isToken = currency.address && currency.address !== Utils.address0x;
-      const bestInterestRate = this.getBestInterest(symbol);
-      const formattedCurrency: CurrencyItem = {
-        address,
-        symbol,
-        img,
-        isToken,
-        bestInterestRate
-      };
-      currencies.push(formattedCurrency);
-    });
-
-    this.currencies = currencies;
+  constructor(
+    private chainService: ChainService
+  ) {
+    this.buildCurrencies();
   }
 
   /**
@@ -96,6 +77,20 @@ export class CurrenciesService {
   }
 
   /**
+   * Return only this currencies
+   * @return Currencies
+   */
+  getCurrenciesByKey(key: 'symbol' | 'address', value: string | string[]): Array<CurrencyItem> {
+    let repulsedValues = typeof value === 'string' ? [value] : value;
+    repulsedValues = repulsedValues.map((item) => item.toLowerCase());
+
+    const filteredCurrency: Array<CurrencyItem> = this.currencies.filter(
+      currency => repulsedValues.includes(currency[key].toLowerCase())
+    );
+    return filteredCurrency;
+  }
+
+  /**
    * Return all currencies with a exception
    * @return Currencies
    */
@@ -132,4 +127,48 @@ export class CurrenciesService {
     };
   }
 
+  /**
+   * Get currency decimals
+   * @param key CurrencyItem key
+   * @param value Symbol or token address
+   * @return Decimals
+   */
+  getCurrencyDecimals(key: 'symbol' | 'address', value: string): number {
+    const { symbol } = this.getCurrencyByKey(key, value);
+    const { decimals: expectedDecimals } = new Currency(symbol);
+    const { config } = this.chainService;
+    const decimals = config.currencies.currencyDecimals[symbol] || expectedDecimals;
+    return decimals;
+  }
+
+  /**
+   * Build currencies using the current chain config
+   */
+  private buildCurrencies() {
+    const { config } = this.chainService;
+    const envCurrencies = config.currencies.usableCurrencies;
+    const currencies: CurrencyItem[] = [];
+
+    envCurrencies.map((currency: {
+      symbol: string;
+      address: string;
+      img: string;
+    }) => {
+      const address = currency.address;
+      const symbol = currency.symbol;
+      const img = currency.img;
+      const isToken = currency.address && currency.address !== Utils.address0x;
+      const bestInterestRate = this.getBestInterest(symbol);
+      const formattedCurrency: CurrencyItem = {
+        address,
+        symbol,
+        img,
+        isToken,
+        bestInterestRate
+      };
+      currencies.push(formattedCurrency);
+    });
+
+    this.currencies = currencies;
+  }
 }