+++ src/app/services/proxy-api.service.ts
@@ -1,7 +1,8 @@
 import { Injectable } from '@angular/core';
-import { Engine } from './../models/loan.model';
-import { ApiResponse } from './../interfaces/api-response';
-import { ApiService } from './api.service';
+import { Engine } from 'app/models/loan.model';
+import { ApiResponse } from 'app/interfaces/api-response';
+import { ChainService } from 'app/services/chain.service';
+import { ApiService } from 'app/services/api.service';
 
 @Injectable({
   providedIn: 'root'
@@ -9,6 +10,7 @@ import { ApiService } from './api.service';
 export class ProxyApiService {
 
   constructor(
+    private chainService: ChainService,
     private apiService: ApiService
   ) { }
 
@@ -21,9 +23,13 @@ export class ProxyApiService {
    * @return Loans array
    */
   async getRequests(page = 1, pageSize = 20, sort?: object, filters?: object): Promise<ApiResponse> {
+    const { isEthereum } = this.chainService;
+    if (!isEthereum) {
+      return await this.apiService.getRequests(Engine.UsdcEngine, page, pageSize, sort, filters).toPromise();
+    }
+
     const usdcEngine = await this.apiService.getRequests(Engine.UsdcEngine, page, pageSize, sort, filters).toPromise();
     const rcnEngine = await this.apiService.getRequests(Engine.RcnEngine, page, pageSize, sort, filters).toPromise();
-
     const allResults: ApiResponse = {
       content: usdcEngine.content.concat(rcnEngine.content),
       meta: {
@@ -44,9 +50,13 @@ export class ProxyApiService {
    * @return Loans array
    */
   async getBorrowed(account: string, page = 1, pageSize = 20, sort?: object): Promise<ApiResponse> {
+    const { isEthereum } = this.chainService;
+    if (!isEthereum) {
+      return await this.apiService.getBorrowed(Engine.UsdcEngine, account, page, pageSize, sort).toPromise();
+    }
+
     const usdcEngine = await this.apiService.getBorrowed(Engine.UsdcEngine, account, page, pageSize, sort).toPromise();
     const rcnEngine = await this.apiService.getBorrowed(Engine.RcnEngine, account, page, pageSize, sort).toPromise();
-
     const allResults: ApiResponse = {
       content: usdcEngine.content.concat(rcnEngine.content),
       meta: {
@@ -67,9 +77,13 @@ export class ProxyApiService {
    * @return Loans array
    */
   async getLent(account: string, page = 1, pageSize = 200, sort?: object): Promise<ApiResponse> {
+    const { isEthereum } = this.chainService;
+    if (!isEthereum) {
+      return await this.apiService.getLent(Engine.UsdcEngine, account, page, pageSize, sort).toPromise();
+    }
+
     const usdcEngine = await this.apiService.getLent(Engine.UsdcEngine, account, page, pageSize, sort).toPromise();
     const rcnEngine = await this.apiService.getLent(Engine.RcnEngine, account, page, pageSize, sort).toPromise();
-
     const allResults: ApiResponse = {
       content: usdcEngine.content.concat(rcnEngine.content),
       meta: {
@@ -90,9 +104,13 @@ export class ProxyApiService {
    * @return Loans array
    */
   async getAcvivity(page = 1, pageSize = 20, sort?: object, filters?: object): Promise<ApiResponse> {
+    const { isEthereum } = this.chainService;
+    if (!isEthereum) {
+      return await this.apiService.getAcvivity(Engine.UsdcEngine, page, pageSize, sort, filters).toPromise();
+    }
+
     const usdcEngine = await this.apiService.getAcvivity(Engine.UsdcEngine, page, pageSize, sort, filters).toPromise();
     const rcnEngine = await this.apiService.getAcvivity(Engine.RcnEngine, page, pageSize, sort, filters).toPromise();
-
     const allResults: ApiResponse = {
       content: usdcEngine.content.concat(rcnEngine.content),
       meta: {
@@ -122,13 +140,16 @@ export class ProxyApiService {
    * @return Last and current block
    */
   async getApiStatus(): Promise<{last_block: number, current_block: number}> {
+    const { isEthereum } = this.chainService;
+    if (!isEthereum) {
+      return await this.apiService.getApiStatus(Engine.UsdcEngine).toPromise();
+    }
+
     const usdcEngine = await this.apiService.getApiStatus(Engine.UsdcEngine).toPromise();
     const rcnEngine = await this.apiService.getApiStatus(Engine.RcnEngine).toPromise();
-
     if (usdcEngine.last_block < rcnEngine.last_block) {
       return usdcEngine;
     }
-
     return rcnEngine;
   }
 }